requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0  # Used by yfinance and for data manipulation

# Data providers - Equities
massive>=1.0.0  # Massive.com (formerly Polygon.io) - Recommended for production
yfinance>=0.2.0  # Yahoo Finance - Free alternative for development

# Data providers - Crypto
python-binance>=1.0.0  # Binance API

# Data providers - Forex
oandapyV20>=0.6.3  # OANDA v20 API

# Database drivers (install as needed based on your database choice)
# PostgreSQL
psycopg2-binary>=2.9.0

# MySQL (optional)
# pymysql>=1.0.0

# SQLite (built-in with Python, no package needed)

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
